# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: solar
    description: Solar energy monitoring operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /solar/history:
    get:
      operationId: getSolarHistory
      tags: [solar]
      summary: Get solar energy history
      description: Returns the latest consumption, production, balance and time labels from the connected solar API.
      responses:
        "200":
          description: Solar history
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SolarHistory"
        "502":
          description: Upstream solar API unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    SolarHistory:
      type: object
      properties:
        consommation:
          type: array
          items:
            type: number
        production:
          type: array
          items:
            type: number
        solde:
          type: array
          items:
            type: number
        labels:
          type: array
          items:
            type: string
      required:
        - consommation
        - production
        - solde
        - labels
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message