# lib/api-zod/src/generated/types/solarHistory.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface SolarHistory {
  consommation: number[];
  production: number[];
  solde: number[];
  labels: string[];
}